% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_hawaii.R
\name{check_endemic_hawaii}
\alias{check_endemic_hawaii}
\title{Check Hawaii endemicity}
\usage{
check_endemic_hawaii(
  hawaii_land_features = list(),
  buffer_distance_km = 50,
  region_demarcation_points = list(main_islands = c(-160.1, 21.8), nwhi = c(-161.9,
    23), more_west = c(-172, 25)),
  species_range
)
}
\arguments{
\item{hawaii_land_features}{A list of (named) sf-objects or spatial
data representations of land features of Hawaii.}

\item{buffer_distance_km}{A buffer distance in kilometres. How far off the
coast of Hawaii land features should species be considered to occur 'in/on/
off Hawaii'?}

\item{region_demarcation_points}{A list of 2-element
vectors which will be used to construct Voronoi polygons from the
Hawaii buffer. Set to Niihau (main_islands) and Nihoa (NWHI).}

\item{species_range}{A species range as sf-POLYGON object
or spatial data file names.}
}
\value{
A data frame of the species global range in Hawaii.
}
\description{
Check Hawaii endemicity
}
