% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_area_data.R
\name{end_check_area}
\alias{end_check_area}
\title{Check and read data.}
\usage{
end_check_area(spatial_data, check_polygon = TRUE)
}
\arguments{
\item{spatial_data}{A spatial representation of the area to read.
May be any GDAL compatible file, or an \code{sf}-*POLYGON.
Assumed to be in the WGS84 system (normal long-lat).}

\item{check_polygon}{Whether to check for POLYGON type and break if not.}
}
\value{
An sf polygon object.
}
\description{
Read in spatial data from a local file of any GDAL supported type.
}
