% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_endemic.R
\name{end_check_endemic}
\alias{end_check_endemic}
\title{Check endemicity based on species range.}
\usage{
end_check_endemic(aoi, utm_epsg_code, buffer_distance_km, sp_range)
}
\arguments{
\item{aoi}{A spatial representation of the area in which the
species is expected to be endemic. Must be an
\code{sf}-*POLYGON. Assumed to be in the WGS84 system (normal long-lat).}

\item{utm_epsg_code}{An EPSG code (see e.g. epsg.io/4326) for the UTM system
appropriate to the area of interest. Necessary to draw buffers in metres when
the coordinates are provided in longitude-latitude.}

\item{buffer_distance_km}{A buffer distance in kilometres. Primarily
useful for marine species which may be considered endemic to waters off a
feature such as an island.}

\item{sp_range}{A spatial representation of the species' range.
May be any GDAL compatible file, or an \code{sf}-*POLYGON.
Assumed to be in the WGS84 system (normal long-lat).}
}
\value{
The proportion of the species range within a buffer around the area
of interest.
}
\description{
Check endemicity based on species range.
}
