% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_voronoi_area.R
\name{demarcate_regions}
\alias{demarcate_regions}
\title{Divide an area into Voronoi polygons}
\usage{
demarcate_regions(
  area_of_interest,
  region_demarcation_points = list(main_islands = c(-160.1, 21.8), nwhi = c(-161.9,
    23))
)
}
\arguments{
\item{area_of_interest}{An sf-object showing the area of interest.
Must have a CRS. Voronoi points will be converted to this CRS.}

\item{region_demarcation_points}{A list of 2-element
vectors which will be used to construct Voronoi polygons from the
Hawaii buffer. Set to Niihau (main_islands) and Nihoa (NWHI).
Must be specified in decimal degrees (WGS84 CRS: EPSG code = 4326).}
}
\value{
An sf-object with the area of interest demarcated into regions.
}
\description{
Divide an area into Voronoi polygons
}
